code review two
* the maintenance functions are placed in a separate file
